# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LeuwongRR LLM Gateway
  version: 1.0.0
  description: >-
    Product surface in front of OmniRoute. Every path below is exact-matched
    against the route allowlist in src/policy/allowlist.ts; there is no
    catch-all passthrough, so an undocumented path is refused with 404 before
    authentication runs. tests/openapi-contract.test.ts fails the build if this
    document and the allowlist disagree.
servers:
  - url: https://api.leuwongrr.cloud
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: Tenant API key issued by the console.
    internalReadyToken:
      type: apiKey
      in: header
      name: x-internal-ready-token
    internalMetricsToken:
      type: apiKey
      in: header
      name: x-internal-metrics-token
    sessionCookie:
      type: apiKey
      in: cookie
      name: lwrr_session
    accessAssertion:
      type: apiKey
      in: header
      name: cf-access-jwt-assertion
paths:
...
  /v1/models:
    get:
      summary: List models the calling tenant is entitled to
      security:
        - apiKey: []
      responses:
        '200':
          description: Entitled models only; an unentitled tenant receives an empty list.
        '401':
          description: Missing or invalid API key.
        '403':
          description: Key lacks the models:read scope.
  /v1/chat/completions:
    post:
      summary: OpenAI chat completions
      security:
        - apiKey: []
      responses:
        '200':
          description: Completion, or an SSE stream when stream is true.
        '400':
          description: Body failed schema validation.
        '403':
          description: Scope, entitlement, or funding refused the request.
        '429':
          description: Rate limited.
  /v1/responses:
    post:
      summary: OpenAI responses
      security:
        - apiKey: []
      responses:
        '200':
          description: Response, or an SSE stream when stream is true.
        '400':
          description: Body failed schema validation.
        '403':
          description: Scope, entitlement, or funding refused the request.
  /v1/messages:
    post:
      summary: Anthropic messages
      security:
        - apiKey: []
      responses:
        '200':
          description: Message, or an SSE stream when stream is true.
        '400':
          description: Body failed schema validation.
        '404':
          description: Unknown model, in the Anthropic error envelope.
  /v1/messages/count_tokens:
    post:
      summary: Anthropic token count
      security:
        - apiKey: []
      responses:
        '200':
          description: Input token count.
        '400':
          description: Body failed schema validation.
...